strimzi_k8s_faust/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "POC of Strimzi in k8s, using faust and FastAPI"
authors = ["Alejandro Amar Gil <alejandro.amar@sourcemeridian.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
# python = ">3.9, <3.12"
python = "^3.12"

fastapi = { extras = ["standard"], version = "^0.115.5" }
mypy = "^1.13.0"
ruff = "^0.8.0"
cython = "^3.0.11"
six = "^1.16.0"
kafka-python = "1.4.7"
redis = "^5.2.0"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
aiokafka = "^0.12.0"
setuptools = "^75.6.0"
faust-streaming = "^0.11.3"
locust = "^2.32.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
    "F811", # Redefinition of unused function pytest
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]


[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
indent-style = "space"
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.mypy]
python_version = ">3.9, <3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = '^kombu/'


[[tool.mypy.overrides]]
module = ["src.*"]
warn_unused_ignores = false
ignore_missing_imports = true
